Replace color theme context with bgColorService
Replace the color theme context which wrapped all pages to provide
a global color them with bgColorService module which maintains state
outside of the React app.

The reason for the change is to fix the problem where the browser
default color picker would be removed from the dom whenever state
was updated. This behavior prevented users from having the ability
to drag the color picker around on the color well while viewing the
updated background in real-time.