Add next and previous quote buttons
Add QuoteNav module which exports components to render buttons for
navigating to the previous/next quotes which aren't visible unless
the mouse hovers over them. Each button is on the left or right side
of the window using fixed positioning.

Add swipe touch gesture for left/right to change to prev/next quote.